+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2025-XX-XX
+### Added
+- Add Selection Mask Alternative Action
+- Add Create Selection Mask Popup Window Action
 
 ## 2025-12-02
 ### Added